Refactor to improve readability
- Introduce `SensorFrameCreator` to define `create_3dpointcloud_frame()`
  and  `create_image_frame()` functions
- Use `SensorFrameCreator` as class member of `SegmentsSampleCreator`
- Merge `image_setup.py` and `pcd_setup.py` into `sample_formats.py`
- Remove `pcd_setup.py` and rename `image_setup.py` into
  `sample_formats.py`
